requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
motor==3.3.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.1.0.post1 